APEXMALHAR-2220 Move the FunctionOperator to Malhar library
