+++ stream/src/main/java/org/apache/apex/malhar/stream/api/ApexStream.java
@@ -22,9 +22,9 @@ import java.util.concurrent.Callable;
 
 import org.joda.time.Duration;
 
+import org.apache.apex.malhar.lib.function.Function;
 import org.apache.apex.malhar.lib.window.TriggerOption;
 import org.apache.apex.malhar.lib.window.WindowOption;
-import org.apache.apex.malhar.stream.api.function.Function;
 import org.apache.hadoop.classification.InterfaceStability;
 
 import com.datatorrent.api.Attribute;
